Updated Tags view and Edit Metadata view. Added Language property to documents
Tags view: Adding tags works, view is correctly updated
Deleting tags doesn't work quite yet. Wierd bug.
- Have at least 4 tags
- Delete the second tag (this works as expected)
- Again delete the second tag (of the 3 left)
Now the tag cannot be deleted (doesn't exist?) altough it seams the
right name of the tag is passed to the delete_tag function...

Metadata view: Mostly works, only when a date is picked via the
date-picker, it's date is shown in the <input> field, but the data
in the model "OriginalDateString" isn't updated. When the user
manually types a date, these changes are reflected in the model.